Change all Int to Int64 to keep type consistency.

+++ src/samplers/is.jl
@@ -1,5 +1,5 @@
 immutable IS <: InferenceAlgorithm
-  n_samples :: Int
+  n_samples :: Int64
 end
 
 type ImportanceSampler{IS} <: Sampler{IS}
